To-html: Add text rule. Don't use groups ('block', 'inline')
