+++ src/html/blocks/html.js
@@ -1,18 +1,5 @@
-const { Serializer, Deserializer } = require('../../');
+const { Deserializer } = require('../../');
 const parse = require('./parse');
-
-/**
- * Default rule to serialize to HTML. Should be removed in the end.
- * @type {Serializer}
- */
-const serialize = Serializer()
-.then((state) => {
-    // Match any node
-    const node = state.peek();
-    return state
-        .shift()
-        .write(node.text);
-});
 
 /**
  * Deserialize an HTML string
@@ -26,5 +13,5 @@ const deserialize = Deserializer()
         .skip(state.text.length);
 });
 
-module.exports = { serialize, deserialize };
+module.exports = { deserialize };
