+++ src/html/index.js
@@ -1,3 +1,8 @@
 const block = require('./blocks');
+const inline = require('./inlines');
 
-module.exports = { block };
+const ALL = block.concat(inline);
+
+// We always use all rule at once, since we deserialize using only one
+// general rule, and the serialization has no ambiguity.
+module.exports = { block: ALL };
