+++ src/html/inlines/index.js
@@ -1,0 +1,7 @@
+const text = require('./text');
+const html = require('./html');
+
+module.exports = [
+    text,
+    html
+];
