+++ test/to-html/text/input.yaml
@@ -1,6 +1,6 @@
 nodes:
   - kind: block
-    type: line
+    type: paragraph
     nodes:
       - kind: text
         text: Hello World
