+++ test/to-html/text/output.html
@@ -1,1 +1,1 @@
-Hello World
+<p>Hello World</p>
